Get Varying Colors to Display on Map
First, the number of bins was wrong in the code. The code had 4 bins when the data has 5. The bins are defined in metaData.json and look like: [0.048,0.08,0.103,0.126,0.154,0.195] where the first and last values are the minimum and maximum values of the set and the other values are the boundaries between the bins.

Second, we were passing hardcoded bin boundary values to extractCountyAndMetricDictionary() instead of the dynamic values from the data.