Get Color to Behave on the Map (#104)
* Update Chokidar

I saw this error when running npm start:

Error: No version of chokidar is available. Tried chokidar@2 and chokidar@3.
You could try to manually install any chokidar version.
chokidar@3: Error: Cannot find module 'chokidar'

chokidar is some dependency of React that somehow got out of date. https://www.npmjs.com/package/chokidar

That error lead me to this page: https://stackoverflow.com/a/67739231 which had the solution of running:

Ran npm update chokidar

* Get Varying Colors to Display on Map

First, the number of bins was wrong in the code. The code had 4 bins when the data has 5. The bins are defined in metaData.json and look like: [0.048,0.08,0.103,0.126,0.154,0.195] where the first and last values are the minimum and maximum values of the set and the other values are the boundaries between the bins.

Second, we were passing hardcoded bin boundary values to extractCountyAndMetricDictionary() instead of the dynamic values from the data.

* Clean Up Unused Code

* Get Map Key To Match County Color Display

CountyLayer.js was showing default color if no extra data feature is selected but MapColorLegend.js was showing the error color in that case.